Removed VideoAdvertAgent and moved advert js to common bootstrap
