+++ common/app/assets/javascripts/bootstraps/video.js
@@ -1,32 +1,14 @@
 define([
     "common",
     "modules/detect",
-    "modules/analytics/video",
-    "modules/adverts/video"
+    "modules/analytics/video"
 ], function(
     common,
     detect,
-    Videostream,
-    Advert
+    Videostream
 ) {
 
     var modules = {
-
-        initAdverts: function(config) {
-            common.mediator.on('page:video:ready', function(config, context) {
-                if(!config.page.blockAds) {
-                    var support = detect.getVideoFormatSupport();
-                    var a = new Advert({
-                        el: context.querySelector('.player video'),
-                        support: support,
-                        config: config,
-                        context: context
-                    }).init(config.page);
-                } else {
-                    common.mediator.emit("video:ads:finished", config, context);
-                }
-            });
-        },
 
         initAnalytics: function () {
             common.mediator.on('video:ads:finished', function(config, context) {
@@ -47,7 +29,6 @@ define([
             this.initialised = true;
             common.lazyLoadCss('video', config);
             modules.initAnalytics();
-            modules.initAdverts();
         }
         common.mediator.emit("page:video:ready", config, context);
     };
