+++ dev-build/app/Global.scala
@@ -2,4 +2,4 @@ import conf.RequestMeasurementMetrics
 import controllers.front.FrontLifecycle
 import play.api.mvc.WithFilters
 
-object Global extends WithFilters(RequestMeasurementMetrics.asFilters: _*) with VideoAdLifecycle with MostPopularLifecycle with FrontLifecycle with StoryLifecycle
+object Global extends WithFilters(RequestMeasurementMetrics.asFilters: _*) with MostPopularLifecycle with FrontLifecycle with StoryLifecycle
