+++ video/app/controllers/VideoController.scala
@@ -7,7 +7,7 @@ import play.api.mvc.{ Content => _, _ }
 import play.api.libs.json.Json._
 import play.api.libs.json.JsObject
 
-case class VideoPage(video: Video, storyPackage: List[Trail], advert: Option[JsObject])
+case class VideoPage(video: Video, storyPackage: List[Trail])
 
 object VideoController extends Controller with Logging with ExecutionContexts {
 
@@ -22,15 +22,6 @@ object VideoController extends Controller with Logging with ExecutionContexts {
     }
   }
 
-  private def AdvertToJson(advert: Option[VideoAdvert]): Option[JsObject] = {
-    advert.map{ ad =>
-      toJson(Map(
-        "file" -> toJson(ad.media),
-        "trackingEvents" -> toJson(ad.tracking)
-      )).as[JsObject]
-    }
-  }
-
   private def lookup(path: String)(implicit request: RequestHeader) = {
     val edition = Edition(request)
     log.info(s"Fetching video: $path for edition $edition")
@@ -42,7 +33,7 @@ object VideoController extends Controller with Logging with ExecutionContexts {
         val videoOption = response.content.filter { _.isVideo } map { new Video(_) }
         val storyPackage = response.storyPackage map { new Content(_) }
 
-        val model = videoOption map { video => VideoPage(video, storyPackage.filterNot(_.id == video.id), AdvertToJson(VideoAdvertAgent())) }
+        val model = videoOption map { video => VideoPage(video, storyPackage.filterNot(_.id == video.id)) }
         ModelOrResult(model, response)
     }.recover{suppressApiNotFound}
   }
