+++ video/app/views/fragments/videoBody.scala.html
@@ -14,13 +14,6 @@
             @fragments.headline(video.headline)
         </header>
 
-        @page.advert.map(_.toString).map{ ad =>
-        <script>
-            guardian.config.page.videoAd = @Html(ad);
-        </script>
-        }
-
-
         <div class="player">
             <video controls="controls" poster="@video.imageOfWidth(640).map{ image => @image.path }.getOrElse("")">
                 @video.encodings.map{ encoding =>
